% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{link_pubmed}
\alias{link_pubmed}
\title{Create link to PubMed entry from PMID}
\usage{
link_pubmed(pmid, format = c("markdown", "html"))
}
\arguments{
\item{pmid}{A vector of PMIDs}

\item{format}{Format of links to be created.}
}
\description{
Create link to PubMed entry from PMID
}
