% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{printMD.geneList}
\alias{printMD.geneList}
\title{Print table of differential expression results}
\usage{
\method{printMD}{geneList}(x, ...)
}
\arguments{
\item{x}{An object of class \code{geneList}}

\item{...}{Ignored}
}
\description{
Print table of differential expression results
}
