% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{printMD.topGO}
\alias{printMD.topGO}
\title{Print table of enriched GO terms}
\usage{
\method{printMD}{topGO}(x, de, go_map, symbol_map, ...)
}
\arguments{
\item{x}{An object of class topGO.}

\item{de}{List of entrez IDs for differentially expressed genes.}

\item{go_map}{Environment mapping go terms to entrez IDS.}

\item{symbol_map}{Environment mapping entrez IDs to gene symbols.}

\item{...}{Ignored}
}
\description{
Links to GO terms on AmiGO
}
