% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{printMD.topKEGG}
\alias{printMD.topKEGG}
\title{Print table of KEGG pathways}
\usage{
\method{printMD}{topKEGG}(x, convert = FALSE, genes = NULL, map = NULL,
  annot = NULL, tooltip = c("up", "down"), ...)
}
\arguments{
\item{x}{An object of class \code{topKEGG}}

\item{convert}{LOgical indicating whether gene IDs should be converted.}

\item{genes}{Vector of entrez IDs to highlight.}

\item{map}{An environment mapping entrez IDs to gene symbols.}

\item{tooltip}{Column to use for tooltips.}

\item{...}{Additional arguments are passed to \code{pander}.}

\item{A}{\code{data.frame} with kegg annotations to use. If this is \code{NULL}
annotations will be downloaded from KEGG.}
}
\description{
Links to corresponding entries on the KEGG website are added to the output.
}
